+++ app/src/main/java/com/kabouzeid/gramophone/model/AbsCustomPlaylist.java
@@ -12,7 +12,7 @@ import java.util.List;
  */
 
 public abstract class AbsCustomPlaylist extends Playlist {
-    public AbsCustomPlaylist(int id, String name) {
+    public AbsCustomPlaylist(long id, String name) {
         super(id, name);
     }
