+++ app/src/main/java/com/kabouzeid/gramophone/model/Playlist.java
@@ -8,10 +8,10 @@ import androidx.annotation.NonNull;
  * @author Karim Abou Zeid (kabouzeid)
  */
 public class Playlist implements Parcelable {
-    public final int id;
+    public final long id;
     public final String name;
 
-    public Playlist(final int id, final String name) {
+    public Playlist(final long id, final String name) {
         this.id = id;
         this.name = name;
     }
@@ -35,7 +35,7 @@ public class Playlist implements Parcelable {
 
     @Override
     public int hashCode() {
-        int result = id;
+        int result = (int)id;
         result = 31 * result + (name != null ? name.hashCode() : 0);
         return result;
     }
@@ -56,12 +56,12 @@ public class Playlist implements Parcelable {
 
     @Override
     public void writeToParcel(Parcel dest, int flags) {
-        dest.writeInt(this.id);
+        dest.writeLong(this.id);
         dest.writeString(this.name);
     }
 
     protected Playlist(Parcel in) {
-        this.id = in.readInt();
+        this.id = in.readLong();
         this.name = in.readString();
     }
